Cam header list change
Each camera is maintained on it's own list for easier purging and manipulating.

This means there are 2 camera record lists coming off of each header record (one header record per object).  Each list is time ordered with the newest records at the start of the list so if a purge is done, everything beyond the first "old" record found can be purged.