Fixedd bug in purge thread causing crashes
Problem was if all of the camera records for a specific object ID were purged, the pointer in the  header record was still maintained.  Now we check to see if we are at the top and handle that as a special case.

Tested with over 3000 camera messages processed - used to crash after about 30.  I think we'll call this fixed.