Isolated 5vdc from Raspberry pi
Added USB connection to supply 5vdc to the status display panel.

Added note explaining some Raspberry Pi interfacing and USB power considerations.

Added PDF of schematic for easier viewing - this is convenience item - if you are really interested in exactly what the working schematic is, get the right software and look at the schematic file yourself as it's easy to forget to update the PDF version because it's a completely separate manual step.  See the soon to be created README.md for the status-display or check out the README in the ultrasonic ranger section.