Initial check-in for 9dof sensor
9 degrees of freedom (acceleration,  roll, and magnetic orientation)

We're doing floating point to int conversion and some unnecessary string handling in this server which we could eliminate is we changed the server firmware to dump the records we want to see.

If performance becomes an issue, we can off-load this to the 9dof sensor - even if we incur a performance penalty there, it's probably less of a big deal as we'll just slow our reporting rate and not slow the entire autonomous system.